♻️ [Refactoring]: html.test.tsを機能別に分割
- colorsとflexboxと同様の構造に変更
- __tests__ディレクトリを作成
- 機能別に6つのテストファイルに分割:
  - from: HTML型の作成
  - isEmpty: 空判定
  - normalize: 正規化
  - isValid: 妥当性検証
  - sanitize: サニタイズ
  - toHTMLNode: HTMLNode変換
- describeを削除しフラットな構造に変更
- テスト名を日本語に変更
- test.eachを使用してパラメータ化テストを導入
